backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
requests==2.32.3
beautifulsoup4==4.12.3
firebase-admin==6.5.0
python-dotenv==1.0.1
